+++ lib/components.dart
@@ -1,8 +1,9 @@
 library components;
 
+export 'components/animated_list_component.dart';
 export 'components/card_component.dart';
 export 'components/dropdown_component.dart';
-export 'components/filled_button_async_component.dart';
+export 'components/elevated_button_async_component.dart';
 export 'components/indexed_stack_component.dart';
 export 'components/number_input_component.dart';
 export 'components/open_container_card_component.dart';