+++ lib/components/scaffold_form_component.dart
@@ -97,32 +97,17 @@ class _ScaffoldFormComponentState extends State<ScaffoldFormComponent> {
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   crossAxisAlignment: CrossAxisAlignment.center,
                   children: [
-                    TextButton.icon(
-                      style: ButtonStyle(
-                        foregroundColor:
-                            WidgetStateProperty.resolveWith<Color?>(
-                          (Set<WidgetState> states) {
-                            if (states.contains(WidgetState.pressed)) {
-                              return Theme.of(context)
-                                  .colorScheme
-                                  .tertiary
-                                  .withOpacity(0.6);
-                            }
-
-                            return Theme.of(context).colorScheme.tertiary;
-                          },
-                        ),
-                      ),
+                    ElevatedButton.icon(
                       onPressed: () {
                         Navigator.maybePop(context);
                       },
                       icon: const BackButtonIcon(),
                       label: const Text("Voltar"),
                     ),
-                    FilledButtonAsyncComponent(
+                    ElevatedButtonAsyncComponent(
                       icon: Icons.send,
                       label: 'Enviar',
-                      pressedLabel: 'Enviando...',
+                      pressedLabel: ' Enviando...',
                       onPressed: () => widget.handleSubmit(),
                     ),
                   ],