+++ lib/src/product/product_details_controller.dart
@@ -1,7 +1,47 @@
+import 'package:fixnum/fixnum.dart';
 import 'package:flutter/material.dart';
 import 'package:gs_manager/helpers.dart';
+import 'package:gs_manager/protos.dart';
 
 class ProductDetailsController {
+  ProductDetailsController({
+    this.productUpdating,
+  });
+
+  final GetProductByIdResponse? productUpdating;
+  late CreateProductRequest product =
+      //  productUpdating == null?
+      CreateProductRequest(
+    name: "",
+    // pictureBinary: x,
+    productBrandId: null,
+    productCategoryId: null,
+    variants: <ProductVariant>[
+      ProductVariant(
+        barCode: "",
+        sku: "",
+        color: "",
+        size: "",
+        inventory: ProductVariantInventory(
+          minimumStockAmount: 0,
+          quantityAvailable: 0,
+        ),
+        unitPrice: DecimalValue(
+          units: Int64(0),
+          nanos: 0,
+        ),
+      ),
+    ],
+  )
+      // : CreateProductRequest(
+      //     name: "",
+      //     // pictureBinary: x,
+      //     productBrandId: "",
+      //     productCategoryId: "",
+      //     variants: ,
+      //   )
+      ;
+
   final GlobalKey<FormState> formKey = GlobalKey<FormState>();
 
   Future<bool> handleSubmit(BuildContext context) async {